pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-utils"
version = "0.1.0"
authors = [
  { name="Operator ITA", email="info@operator-ita.com" },
]
description = "A clean, unified SDK for multi-provider LLM communication."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "google-genai",
    "openai",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_utils*"]
